HT_15/project/requirements.txt
Scrapy-2.5.1
beautifulsoup4=4.10.0
requests=2.27.1
itemadapter~=0.4.0

